Add files via upload
Update: ScanWindow starts by showing a dialogbox that asks for itemNumber and only takes ints, X or Cancel closes everything. Once input, the itemNumber is run against the db, if it is in the db the user is notified and the textfields are populated with the info from the db, otherwise the new itemNumber is placed in the fields and the user has access to RequiredInfo fields. Added a section to Add to current stock however Errors in the SQL Handler currently prevent it from working, but fails by letting user know db was not updated before closing. ItemType will be addressed next.